+++ Videostop/src/Logic.c
@@ -67,24 +67,24 @@ void L_Restart()
 
 void L_Update()
 {
-  static int before;
-  static int idle_time;
-  static int stop_time;
-  static int statusbar_time;
+  static unsigned before;
+  static unsigned idle_time;
+  static unsigned stop_time;
+  static unsigned statusbar_time;
 
 #if defined(USE_FREEGLUT)
-  int const now = glutGet(GLUT_ELAPSED_TIME);
+  unsigned const now = (unsigned) glutGet(GLUT_ELAPSED_TIME);
 #endif
 
 #if defined(USE_GLFW)
-  int const now = (int) (glfwGetTime() * 1000.0);
+  unsigned const now = (unsigned) (glfwGetTime() * 1000.0);
 #endif
 
 #if defined(USE_SDL2)
-  int const now = SDL_GetTicks();
+  unsigned const now = SDL_GetTicks();
 #endif
 
-  int const elapsed = now - before;
+  unsigned const elapsed = now - before;
   before = now;
 
   if (g_world.dice_count != g_config.dice_count)
